+++ src/components/Mypage/loginInfo.tsx
@@ -1,10 +1,7 @@
 import React, { useEffect, useState } from "react";
-import styled from "styled-components";
 import { useQuery } from "@tanstack/react-query";
-import * as S from "@/styles/login/index.styles";
-import Button from "@/components/common/Button";
-import Input from "@/components/common/Input";
-import useMypage from "@/hooks/mypage/useMypage";
+import { StyledInput } from "@/styles/login/index.styles";
+import useLoginInfo from "@/hooks/mypage/loginInfo/useLoginInfo";
 import QUERYKEYS from "@/constants/querykey";
 import { loadMe } from "@/api/auth";
 import useSignup from "@/hooks/useSignup";
@@ -12,112 +9,12 @@ import LoadingBar from "@/components/LoadingBar";
 import Dialog from "@/components/common/Dialog";
 import PATH from "@/constants/path";
 import { useRouter } from "next/dist/client/router";
+import * as S from "@/styles/mypage/loginInfo/index.styles";
 
-interface ButtonProps {
-  inValid: boolean;
-}
-const Wrapper = styled.div<{ isLoading?: boolean }>`
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-  padding-top: 4rem;
-  //height: 100%;
-  width: 55rem;
-  height: ${({ isLoading }) => (isLoading ? "50rem" : "auto")};
-  @media (max-width: 1007px) {
-    width: 94%;
-  }
-`;
-const Box = styled.div<ButtonProps>`
-  display: flex;
-  flex-direction: row;
-  align-items: ${({ inValid }) => (inValid ? "center" : "flex-end")};
-`;
-const PasswordInput = styled(Input)`
-  width: 95%;
-  height: 4rem;
-  font-size: 1.2rem;
-  &::placeholder {
-    font-size: 1.2rem;
-    border: none;
-    font-weight: 400;
-  }
-  &:focus {
-    border-bottom: 2px solid;
-  }
-  padding-top: 1.2rem;
-  padding-bottom: 1.2rem;
-  @media (max-width: 1007px) {
-    font-size: 1.4rem;
-    &::placeholder {
-      font-size: 1.4rem;
-      border: none;
-      font-weight: 400;
-    }
-  }
-`;
-const PasswordChangeButton = styled(Button)`
-  border-radius: 0.625rem;
-  width: 5.3rem;
-  height: 3rem;
-  padding: 0.62rem;
-  font-size: 1rem;
-`;
-const DeleteButtonWrapper = styled.div`
-  padding-top: 5rem;
-  display: flex;
-  justify-content: flex-end; /* 오른쪽 끝으로 이동 */
-  width: 55.7rem;
-  @media (max-width: 1007px) {
-    width: 95%;
-  }
-`;
-const MobileLogoutButton = styled(Button)`
-  border: none;
-  border-radius: 0;
-  width: 8rem;
-  height: 3rem;
-  padding: 0.62rem;
-  color: red;
-  font-weight: 300;
-  font-size: 1.4rem;
-
-  @media (min-width: 1007px) {
-    display: none;
-  }
-  &:hover {
-    border: none;
-    color: red;
-    font-weight: 500;
-  }
-  &:focus {
-    font-weight: 500;
-  }
-`;
-const DeleteButton = styled(Button)`
-  border: none;
-  border-radius: 0;
-  width: 6rem;
-  height: 3rem;
-  padding: 0.62rem;
-  font-size: 1rem;
-  @media (max-width: 1007px) {
-    width: 8rem;
-    font-size: 1.4rem;
-  }
-  &:hover {
-    border: none;
-    color: red;
-    font-weight: 500;
-  }
-  &:focus {
-    font-weight: 500;
-  }
-`;
 export default function LoginInfo() {
   const { passwordCheck, onChangePasswordCheck, password, onChangePassword } =
     useSignup();
-  const { mutateChangePassword, deleteUser, logout } = useMypage();
+  const { mutateChangePassword, deleteUser, logout } = useLoginInfo();
   const { data, isLoading } = useQuery({
     queryKey: [QUERYKEYS.LOAD_ME],
     queryFn: loadMe,
@@ -160,7 +57,6 @@ export default function LoginInfo() {
   }, [isLoading]);
 
   if (isLoading && !timedOut) return <LoadingBar type={6} />;
-  console.log("/////////////////", isLoading, timedOut);
 
   const inputs = [
     {
@@ -224,10 +120,10 @@ export default function LoginInfo() {
       )}
       <h2>로그인 정보</h2>
       {inputs.map((input) => (
-        <Wrapper isLoading={isLoading}>
+        <S.Wrapper isLoading={isLoading}>
           {input.buttonTitle ? (
-            <Box inValid={input.inValid}>
-              <PasswordInput
+            <S.Box inValid={input.inValid}>
+              <S.PasswordInput
                 key={input.label}
                 placeholder={input.placeholder}
                 label={input.label}
@@ -239,7 +135,7 @@ export default function LoginInfo() {
                 disabled={isSocial}
                 readOnly={isSocial}
               />
-              <PasswordChangeButton
+              <S.PasswordChangeButton
                 title="변경"
                 onClick={() => {
                   mutateChangePassword.mutate({ newPassword: password });
@@ -248,9 +144,9 @@ export default function LoginInfo() {
                   !(password === passwordCheck && passwordCheck.length > 0)
                 }
               />
-            </Box>
+            </S.Box>
           ) : (
-            <S.StyledInput
+            <StyledInput
               key={input.label}
               placeholder={input.placeholder}
               label={input.label}
@@ -261,12 +157,12 @@ export default function LoginInfo() {
               readOnly={input.readOnly}
             />
           )}
-        </Wrapper>
+        </S.Wrapper>
       ))}
-      <DeleteButtonWrapper>
-        <MobileLogoutButton title="로그아웃" onClick={logout} />
-        <DeleteButton title="탈퇴하기" onClick={deleteUser} />
-      </DeleteButtonWrapper>
+      <S.DeleteButtonWrapper>
+        <S.MobileLogoutButton title="로그아웃" onClick={logout} />
+        <S.DeleteButton title="탈퇴하기" onClick={deleteUser} />
+      </S.DeleteButtonWrapper>
     </div>
   );
 }
